+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+- hw-ledger, keyring, util, util-crypto 12.3.1
 - extension-dapp 0.46.4
 - api, api-contract, types 10.8.1
 - react-{identicon, qr}, ui-{keyring, settings}, vue-identicon 3.4.2